WIP: Allow query parameters to be optional
In order to completely skip a certain condition from the query, we
allow some parameter type to be "optional", this can be configure
on a per parameter basis in the frontend. This can then be used
in tandem with mustach sections to skip a certain block of a query
when such a parameter is not set:

SELECT
    *
FROM
    table
WHERE
    true
{{#param}}
    AND col = '{{ param }}'
{{/param}}

Since mustache will evaluate the variable as a boolean, all python
boolean evaluation rules apply, which brings a couple of pitfalls.

 * If you make a QueryBasedDropdown optional and one of the query
   values is an empty string, we condition will fail and the section
   will not be rendered.

 * Right now making a number optional will prevent users from passing
   as 0 for a value, but we could easily work around this by casting
   integers to strings to make it evaluate to true.