WIP: Allow query parameters to be escaped
In order for "readonly users" to be able to use "text" fields, we
provide a datasource dependent, optional method of escaping the
text input in order to prevent any SQL injection and make the input
"safe". This comes with the automatic benefit of being able to
escape the values returned for a QueryBasedDropdown.

Parameters need to be individually marked for escaping, they will not
be quoted though for compatibility reasons.

A datasource needs to support escaping and the code is built in a way
that does allow for a datasource/query_runner instance to open a
connection in case that is required for properly escaping the input.