WIP: Properly handle scheduled queries without a previous execution
Passing "now" for a query that has never been executed before seems to
be the exact opposite of what we'd want to do, because we are pretending
it has been executed very *recently* instead of *never*.

What we probably should do is honor the interval as well as any other
constraints like a specific time for execution (in case of daily or
weekly schedules).

So as a first, naive attempt, we just reset the execution to the start
of the current day. This should set back the time far enough to trigger
immediate execution while also allowing meaningful comparisons with
`time` and `day_of_week` to match against a somewhat sensible point in
time. This of course assumes that it is better to defer execution to the
"correct" point in time in face of a missing query result instead of
fetching it ASAP.